--- a/pkg/db/linked_list/errors.go
+++ b/pkg/db/linked_list/errors.go
@@ -1,0 +1,7 @@
+package linkedlist
+
+import "errors"
+
+var (
+	ErrNotFound = errors.New("an entry with such a key was not found")
+)