## Before: 
goos: darwin
goarch: arm64
pkg: github.com/PepsiKingIV/KeyValueDB/pkg/db
### BenchmarkHash
BenchmarkHash-8   	 9390354	       113.7 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	github.com/PepsiKingIV/KeyValueDB/pkg/db	1.430s
### BenchmarkPut
BenchmarkPut-8             10000            299037 ns/op            3767 B/op          0 allocs/op
PASS
ok      github.com/PepsiKingIV/KeyValueDB/pkg/db        3.206s
### BenchmarkGet
BenchmarkGet-8           3013459               393.9 ns/op             0 B/op          0 allocs/op
PASS
ok      github.com/PepsiKingIV/KeyValueDB/pkg/db        1.805s
### BenchmarkDelete
BenchmarkDelete-8        2994486               380.2 ns/op             0 B/op          0 allocs/op
PASS
ok      github.com/PepsiKingIV/KeyValueDB/pkg/db        1.762s

## Now:
oos: darwin
goarch: arm64
pkg: github.com/PepsiKingIV/KeyValueDB/pkg/db
=== RUN   BenchmarkHash
### BenchmarkHash
BenchmarkHash-8          9930613               114.2 ns/op             0 B/op          0 allocs/op
PASS
ok      github.com/PepsiKingIV/KeyValueDB/pkg/db        1.475s
=== RUN   BenchmarkPut
### BenchmarkPut
BenchmarkPut-8            959034              1048 ns/op             960 B/op         20 allocs/op
PASS
ok      github.com/PepsiKingIV/KeyValueDB/pkg/db        1.243s
=== RUN   BenchmarkGet
### BenchmarkGet
BenchmarkGet-8           7169835               152.7 ns/op             0 B/op          0 allocs/op
PASS
ok      github.com/PepsiKingIV/KeyValueDB/pkg/db        1.489s
### BenchmarkDelete
BenchmarkDelete-8        6491792               171.8 ns/op             0 B/op          0 allocs/op
PASS
ok      github.com/PepsiKingIV/KeyValueDB/pkg/db        1.521s